Create module for user profile
