+++ client/src/store/module-types.js
@@ -1,2 +1,2 @@
-export const INCREMENT_COUNTER = 'INCREMENT_COUNTER';
-export const DECREMENT_COUNTER = 'DECREMENT_COUNTER';
+export const SET_PROFILE_DATA = 'SET_PROFILE_DATA';
+export const BEST_THING_EVER = 'EMPANADAS';
